requirements.txt
langchain>=0.1.14
langchain-community>=0.0.8
langchain-openai>=0.1.3
langchain-groq
openai>=1.30.1
python-dotenv>=1.0.1
tiktoken>=0.6.0
faiss-cpu>=1.7.4
unstructured>=0.11.6
PyMuPDF>=1.23.24
